requirements.txt
apache-airflow==2.1.2
apache-airflow-providers-amazon==2.0.0
boto3==1.15